requirements.txt
dnspython == 2.3.0
httpx == 0.23.3
redis == 4.5.1